API fix
API fix (overwrite issue)

+++ app/Http/Controllers/Api/AuthenticateController.php
@@ -3,17 +3,14 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use App\Models\APIPasswordResetToken;
-use App\Models\ApiPasswordResetToken as ModelsApiPasswordResetToken;
+use App\Models\ApiPasswordResetToken;
 use App\Models\User;
 use App\Notifications\ApiPasswordResetNotification;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 use Laravel\Sanctum\PersonalAccessToken;
-use Illuminate\Support\Str;
 use Throwable;
 
 class AuthenticateController extends Controller
@@ -69,6 +66,43 @@ class AuthenticateController extends Controller
                 'status' => false,
                 'message' => $th->getMessage(),
             ], 500);
+        }
+    }
+
+    public function unauthorized(Request $request){
+        return response()->json([
+            'status' => false,
+            'message' => 'Unverified user. please login first',
+        ], 401);
+    }
+
+    public function changePassword(Request $request){
+        $validator = Validator::make($request->all(), [
+            'current_password' => 'required',
+            'new_password' => 'required|string|min:8|max:45|confirmed|regex:/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[$#@!%?*-+]).+$/',
+        ]);
+        if ($validator->fails()) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Validation error',
+                'errors' => $validator->errors(),
+            ], 422);
+        }
+
+        $user = auth()->guard('sanctum')->user();
+        if (Hash::check($request->current_password, $user->password)) {
+            $user->password = bcrypt($request->new_password);
+            $user->save();
+
+            return response()->json([
+                'status' => true,
+                'message' => 'Password changed successfully'
+            ], 200);
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'Current password is incorrect'
+            ], 422);
         }
     }
 
@@ -154,14 +188,30 @@ class AuthenticateController extends Controller
 
     public function logout(Request $request)
     {
-        $token = $request->token;
-        $user = Auth()->user();
+        $token = $request->bearerToken();
+        $user = auth()->guard('sanctum')->user();
         $personalAccessToken = PersonalAccessToken::findToken($token);
-        if(
-            $user->id == $personalAccessToken->tokenable_id
-            && get_class($user) == $personalAccessToken->tokenable_type
-        ){
-            $personalAccessToken->delete();
+
+        try{
+            if ($personalAccessToken){
+                if(
+                    $user->id == $personalAccessToken->tokenable_id
+                    && get_class($user) == $personalAccessToken->tokenable_type
+                ){
+                    $personalAccessToken->delete();
+                }
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'Invalid or expired token, please login again',
+                ], 422);
+            }
+
+        }catch(Throwable $th){
+            return response()->json([
+                'status' => false,
+                'message' => $th->getMessage(),
+            ], 500);
         }
 
         return [
@@ -205,8 +255,8 @@ class AuthenticateController extends Controller
             ], 422);
         }
 
-        $reset_token = $resetToken->getResetIdentifierCode();
-       
+        $reset_token = $resetToken->getResetIdentifierCode($resetToken->user_id);
+
         return response()->json([
             'status' => true,
             'token' => $reset_token
@@ -246,12 +296,12 @@ class AuthenticateController extends Controller
         }
 
         $user = User::where('email', $request->email)->first();
-      
+
         do {
-            $obj = new APIPasswordResetToken();
+            $obj = new ApiPasswordResetToken();
             $token = $obj->getResetCode();
             $signature = hash('md5', $token);
-            $exists = APIPasswordResetToken::where([
+            $exists = ApiPasswordResetToken::where([
                 'user_id' => $user->id,
                 'token_signature' => $signature
             ])->exists();
@@ -259,7 +309,7 @@ class AuthenticateController extends Controller
 
         try {
             $user->notify(new ApiPasswordResetNotification($token));
-           
+
             $obj->user_id = $user->id;
             $obj->token_signature = $signature;
             $obj->expires_at = Carbon::now()->addMinutes(30);
@@ -305,29 +355,29 @@ class AuthenticateController extends Controller
         if($verifyToken == null || $verifyToken->count() <= 0 ){
             return response()->json([
                 'status' => false,
-                'message' => 'Invalid token for resetting password'                
+                'message' => 'Invalid token for resetting password'
             ], 422);
         }
 
         $user_id = $verifyToken[0]->user_id;
-        $userInfo = User::where('id', $user_id)->first();       
+        $userInfo = User::where('id', $user_id)->first();
 
         if($userInfo == null || $userInfo->count() <= 0){
             return response()->json([
                 'status' => false,
-                'message' => 'Token does not correspond to any existing user.'                
+                'message' => 'Token does not correspond to any existing user.'
             ], 422);
-        } else if (Carbon::now()->greaterThan($verifyToken[0]->expires_at)){    
+        } else if (Carbon::now()->greaterThan($verifyToken[0]->expires_at)){
             return response()->json([
                 'status' => false,
-                'message' => 'The reset password token has expired. Please try again.'                
+                'message' => 'The reset password token has expired. Please try again.'
             ], 422);
         }
 
         $userObj = User::findOrFail($user_id);
         $userObj->password = bcrypt($request->password);
         $userObj->update();
-        
+
         $verifyToken[0]->update([
             'expires_at' => Carbon::now()
         ]);
@@ -339,5 +389,4 @@ class AuthenticateController extends Controller
         ]);
     }
 
-  
 }
